{"schema":"decodex/commit/1","summary":"Extract Swift live sampler bridge owner","authority":"manual"}
